+++ EVE.SDK/Contracts/IMemory.cs
@@ -7,7 +7,7 @@
         ushort Pc { get; set; }
         int Ir { get; set; }
         ushort Sp { get; set; }
-        void LoadProgram(byte[] program);
+        void LoadProgram(short[] program);
         dynamic Read(int address);
         void Write(int address, dynamic value);
     }
