+++ EVE.SDK/Models/EVEProgram.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; }
         public int Size { get; set; }
-        public byte[] Data { get; set; }
+        public short[] Data { get; set; }
     }
 }
